Motor Control Implementation
Successfully implemented the PID control system equations in the motor.c file, setup STMStudio to interface with the motor via changing the RPM and the gain of the control mechanisms. Successfully verified that I correctly designed my PCB and connected everything to the right peripherals for operation. Updated the Readme, and added the postlab.